Make soci headers/usage consistent
Changed headers to be consistent so that all includes are
"soci/header-file-name.h"

This is functionally the same as changeset 417ef5cc4d08a4685c005d0f2359fc69010a4439
which was 'the great header change'.

I copied include files into src/core/soci - to enable the build
and did not attempt to change the build process so that there
was a single instance of each header. There is a different
reorg of the source tree in a later version of soci where
headers are moved and pulled from the new location as part
of the build - I did not follow through with that set of
changes so we could minimize our risk of changes here.